Design: 스타일 적용
