+++ src/components/SearchBox/SearchBox.module.css
@@ -1,0 +1,11 @@
+.input {
+  flex-grow: 1;
+  padding: 0.25em;
+  font-size: 1rem;
+  border: 0;
+  outline: 0;
+}
+
+.input::placeholder {
+  font-size: 0.875em;
+}
