+++ src/components/SearchBox/SearchBox.tsx
@@ -1,3 +1,5 @@
+import styles from './SearchBox.module.css';
+
 interface SearchBoxProps {
   value: string;
   onChange: (value: string) => void;
@@ -7,6 +9,8 @@ function SearchBox({ value, onChange }: SearchBoxProps) {
   return (
     <input
       type="text"
+      className={styles.input}
+      placeholder="Search for companies, people, or job titles..."
       value={value}
       onChange={(e) => onChange(e.target.value)}
     />
