--- src/components/SearchBtn.tsx
@@ -1,5 +1,0 @@
-function SearchBtn() {
-  return <button aria-label="Search">🔍</button>
-}
-
-export default SearchBtn;
