+++ src/components/SearchBtn/SearchBtn.module.css
@@ -1,0 +1,5 @@
+.btn {
+  background-color: transparent;
+  border: 0;
+  font-size: 1.125rem;
+}
