+++ src/components/SearchBtn/SearchBtn.tsx
@@ -1,0 +1,7 @@
+import styles from './SearchBtn.module.css';
+
+function SearchBtn() {
+  return <button className={styles.btn} aria-label="Search">🔍</button>
+}
+
+export default SearchBtn;
