+++ src/main.tsx
@@ -1,6 +1,7 @@
 import React from 'react'
 import ReactDOM from 'react-dom/client'
 import App from './App.tsx'
+import "./styles/reset.css"
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
   <React.StrictMode>
